Make compatible with python2.7 for nosetests
range expects an in, not a float, in python2.7

+++ arrangement.py
@@ -13,7 +13,7 @@ class Arrangement:
         self.groups = []
         self.participants = json_arrangement
 
-        numGroups = math.ceil(1.0 * len(self.participants) / num_individuals_per_group)
+        numGroups = int(math.ceil(1.0 * len(self.participants) / num_individuals_per_group))
 
         for i in range(numGroups):
             self.groups.append([])
